move computeMaxDisplayWidth to ImageFormatOptionPlugin
