+++ addons/website/static/src/builder/plugins/background_option/background_option.js
@@ -30,7 +30,4 @@ export class BackgroundOption extends BaseOptionComponent {
         const { showColorFilter } = useBackgroundOption(this.isActiveItem);
         this.showColorFilter = showColorFilter;
     }
-    computeMaxDisplayWidth() {
-        return 1920;
-    }
 }
