+++ addons/website/static/src/builder/plugins/background_option/background_option.xml
@@ -29,7 +29,7 @@
         <BackgroundImageOption/>
         <BackgroundPositionOption/>
         <ImageFilterOption level="2"/>
-        <ImageFormatOption level="2" computeMaxDisplayWidth="this.computeMaxDisplayWidth"/>
+        <ImageFormatOption level="2" computeFrom="'background'"/>
         <!-- Color filter -->
         <BuilderRow t-if="this.showColorFilter()" label.translate="Color Filter" level="2">
             <!-- TODO handle all the attributes -->
