+++ addons/website/static/src/builder/plugins/options/background_option.xml
@@ -29,7 +29,7 @@
     </xpath>
     <xpath expr="//ImageFormatOption" position="replace">
         <BuilderContext applyTo="this.websiteBgOptionDomState.applyTo">
-            <ImageFormatOption level="2" computeMaxDisplayWidth="this.computeMaxDisplayWidth"/>
+            <ImageFormatOption level="2" computeFrom="'background'"/>
         </BuilderContext>
     </xpath>
     <xpath expr="//BuilderButton[@action=&quot;'toggleBgImage'&quot;]" position="attributes">
